+++ example/lib/src/app/pages/home/home_controller.dart
@@ -46,14 +46,17 @@ class HomeController extends Controller {
   }
 
   @override
-  void onResumed() {
-    print('On resumed');
-    super.onResumed();
-  }
+  void onResumed() => print('On resumed');
 
   @override
-  void onDisposed(BuildContext context) {
+  void onReassembled() => print('View is about to be reassembled');
+
+  @override
+  void onDeactivated() => print('View is about to be deactivated');
+
+  @override
+  void onDisposed() {
     homePresenter.dispose(); // don't forget to dispose of the presenter
-    super.onDisposed(context);
+    super.onDisposed();
   }
 }